requirements.txt
requests~=2.32.3
tenacity~=8.5.0
pandas
jupyter
httpx<=0.27.0 # (Athina relies on Weaviate which requires httpx<=0.27.0)
athina-client
python-dotenv~=1.0.1
ipython
ipywidgets